% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.DNAextractQC.R
\name{f.DNAextractQC}
\alias{f.DNAextractQC}
\title{f.DNAextractQC}
\usage{
f.DNAextractQC(path = choose.files, LowerOD = 1.65, UpperOD = 2.20, DNAcon = 10, outPath, type, db = NULL)
}
\arguments{
\item{path}{the name of the file which the data are to be read from.  (default: `choose.files`).}

\item{LowerOD}{lower limit of OD260/280 ratio (default: 1.65).}

\item{UpperOD}{upper limit of OD260/280 ratio (default: 2.20).}

\item{DNAcon}{DNA concentration (default: 10 µg/ml).}

\item{outPath}{a character string naming a output path.}

\item{type}{*1* : standard format/ *2* : retest format.}

\item{db}{connection handle returned by `RODBC::odbcConnect`. (default: *NULL*).}
}
\description{
f.DNAextractQC
}
