% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCpassList.R
\name{QCpassList}
\alias{QCpassList}
\title{QCpassList}
\usage{
QCpassList(outPath, db, myFirst, myLast)
}
\arguments{
\item{outPath}{a character string naming a output path.}

\item{db}{connection handle returned by `RODBC::odbcConnect`.}

\item{myFirst}{scanning the first work ID barcode on white box.}

\item{myLast}{scanning the last work ID barcode on white box.}
}
\description{
Check pass list after quality control for DNA extraction.
}
