% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSheet.R
\name{SampleSheet}
\alias{SampleSheet}
\title{SampleSheet}
\usage{
SampleSheet(txt= file.choose(), outPath, db = NULL, GT, plt, barcode)
}
\arguments{
\item{txt}{the name of the file which the data are to be read from.
(default: `choose.files`).}

\item{outPath}{a character string naming a output path.}

\item{db}{connection handle returned by `RODBC::odbcConnect`.
(default: *NULL*).}

\item{GT}{machine number.}

\item{plt}{array number.}

\item{barcode}{scanning the barcode on the array.}
}
\description{
Create sample sheet for Gene Titan Array Plate Registration format file.
}
